requirements.txt
python-telegram-bot==20.6
APScheduler==3.10.4
python-dotenv==1.0.1
